Add support for 'tokio-console' (#619)
### Summary:

See screenshot but tokio-console is an external tool
that let's us connect to the running process and
get a list of tokio::tasks running as well as perf
information on the running instance.  Seems super
useful.

To make this work, I added
``` console_subscriber::init();``` to netdebug_init() so
this will work with all of our binaries.  This causes a
running binary to bind() localhost::6669 by default and
then you install the binary with:

cargo install --locked tokio-console

and then run with

```tokio-console```

If there are multiple binaries on the same machine, the default listening port can be
changed with the env TOKIO_CONSOLE_BIND, e.g., :

`TOKIO_CONSOLE_BIND=127.0.0.1:7000 ./binary`

and then tokio-console can point to the new port with 

`tokio-console tcp://127.0.0.1:7000`

 ### Test Plan:

Ran by hand... seems to work.  See screenshot!

![Screenshot from 2024-02-20 10-28-03.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Se6cfQDK8krglOiJ7Bgl/435bde31-ff29-42fb-ad9d-ee7ffe516baf.png)