Use the default System Certs for DB TLS connections (#629)
### Summary:
Our previous way of getting the root certs only worked on Linux
and wasn't working in my Windows dev environment.

So after some googling, it looks like I didn't need to lookup
the certs in this way and if I build the connector a different
way, we just got the default system certs.

FWIW: I was confused b/c the rustls docs all say you need to get
the certs from the system this way (but they have a special crate
to do it).
 ### Test Plan:

Tested by hand and seems to work at least on windows; was able to
connect to the timescaledb server which requires encrypted connections.