Log connections to remote cloud storage
 ### Summary:
We were just dumping connection logs into a json blob
in SQLITE in local storage.  This PR moves them to be
dumped to timescaledb and expands the schema so it's
some data is stored natively.

I've also expanded the production DB for the new table.

Lots of TODOs but this is a big first step.

 ### Test Plan:

Expanded existing test to write a connection measurement and
store it back.