Added a client UUID for storing logs on server
 ### Summary:
Currently randomly generated at start time; TODO(Gregor) to persist it.

I've done the work to plumb what the client gives us back to the
database BUT we should do something to validate it is what they say
it is... TODO.
 ### Test Plan:
Added it to the test; can write and read back.