webserver db - split out ConnectionKey
 ### Summary:
Left over from two diffs ago.  Don't store the
connection key as a json string, store the components.

NOTE: Sql is pedantic and won't let you store a u16 and
if you convert to i16 (which it allows), it won't let
you store an u16 in an "INT" only in a "SMALLINT".

 ### Test Plan:

Existing tests