Log the webserver's measurements to the remotedb (#651)
### Summary:

I didn't understand/forgot (b/c I wrote it) the code path
for ConnectionMeasurements that were generated locally by
the webserver and they were triggering a panic which I hot patched
and then a warning.  Now they are being logged to the same table
with the desktop traffic, but with the addition of a 'source_type'
field.

Per conversation with Gregor, he prefers that we log this as a string
instead of an enum, so I'm doing that.

 ### Test Plan:
Existing tests