Infrastructure to parse JWTs and access Clerk.com's backend infra (#643)
### Summary:
Created a 'UserService' shared instance which wraps API calls to Clerk.com.  Also started to muck with the front-end logic including the <UserButton /> which gives a surprising amount of controls to the user out of the box

TODO: add the REST hooks to convert a JWT from Clerk into a JWT from us *or* try to embed some private metadata into the users's JWT so we can avoid keeping our own user state.

Added 'clerk_rs' and 'jsonwebtoken' crates as they are super useful.

 ### Test Plan:

Added some new cargo tests, but still lots of work to do