Consolidated all secrets into secrets.toml (#667)
### Summary:
Before I add more secrets to the command line, let's have
a single secrets file with all of the confidential stuff
in one place.  I've brought the timescaledb (read and write, users +
pass) and clerk prod and dev secrets into a single TOML file
(NOT CHECKED IN!!) and wrapped it with TOML for ease of parsing.

 ### Test Plan:
Added a trivial test and did some hand testing... need to try more in
prod and need to roll out the new secrets file to topology.netdebug.com
before we upgrade.