Windows postgres embed test speed up (#689)
### Summary:
Starting and running the postgres embed based tests
is ~1 second on Linux and Mac and ~25seconds on Windows.

Part of the problem is getting the right temp directory and allowing
caching, which this diff fixes.  It drops the time on windows down
to 11 seconds... which is still an order of magnitude slower, but
enough to let me sleep at night.

Cool rust portability call: std::env::temp_dir() finds the OS specific
temp directory - new to me!

 ### Test Plan:
This is the test.