A large refactor of http + SQL tests
 ### Summary:
Problem: we were getting an error from our
tests saying "Can't extract auth session. Is `AuthManagerLayer` enabled?"
which was causing me to go crazy.

I fixed this problem (below) and executed a refactor of our tests so
that we now have live working Mocks of our user auth flows without
Clerk.com magic.

The fundamental problem was that our 'console_login' function
took an 'AuthSession' type which was under the covers a
AuthSession<$prod_auth> type and when we called it with
a $mock_auth, it couldn't find an authentication function that matched.

So, I tried a bunch of things and ultimately ended up with changing
the UserService that our $prod_auth scheme (i.e., NetDebugUserBackend)
takes to accept a 'disable_auth_for_testing' flag and when that flag is
set, the jwt string that's passed to us is accepted blindly as the
user's name.

Given that I now had auth working I reenabled the #[ignored] tests
and moved them to the newly created test_http_routes.rs integration
test, which, after some plumbing, seem to work against fake database
data!
 ### Test Plan:

What if I told you this was the test...