Move auth to cookies instead for From
 ### Summary:

As Gregor and I discussed, having /api/login?clerk_jwt=.... means
that someone could accidentally cache the JWT in a web history which
isn't great so we wanted to move to using the JWT that's already
in the __session cookie provided by Clerk.

This got prioritized to now because apparently you can't use a hook
(e.g., like useAuth, which we were depending on) inside a function
like we had assumed.  The whole code was only working because
we were previously getting the auth token and thus accidentally were
never retriggering the auth.

Also removed the (now dated) POST method of /api/login

 ### Test Plan:

Updated existing tests - they pass!