Change REST /api/login method from GET to POST (#735)
### Summary:

After discussion with Gregor, we decided to be safe and
move this to POST to really really prevent any XSS attacks.
POST methods can never be called outside of CORS checks and
since this "modifies state" by issuing a new cookie, it could
be argued that POST is best.

Also it was an easy change :-)

 ### Test Plan:

Re-ran existing tests 'cargo t' and reloaded 'npm run dev' console