Create a 'netdebug_test_init()' safe for all tests (#751)
### Summary:

We currently can't call 'netdebug_init()' inside tests, e.g.,
to enable logging because the logging functions can only be
called once.  So create a special netdebug_test_init() function
that uses std::sync::Once to ensure the call only happens once.

Note that our previous work around to change, e.g., log::Warn!()
to println!() only works for code we control and only when
\#[cfg(test)] is set.  For integration tests with the DB this
doesn't work because (a) there is a bunch of code in pg_embed we
don't control and \#[cfg(test)] is not set.

NOTE: we want a separate function than the standard netdebug_init()
because we want to default to no logging (e.g., 'error' level).

This does imply that we need to manually add 'netdebug_test_init()' to
all of our tests (or at least integration tests) which is obviously
ugly.  TODO to figure out how to make our own test macro but it
looks like a PITA and the test frameworks I've found don't seem
trivial.  Long story short: that's a nice to have but seems like more
work than it's worth for this patch.

 ### Test Plan:

Ran 'cargo t' with two tests with this set and they don't panic!() or
print extra garbage.